introduce widget types
There is quite a lot of shared/common Window methods and it makes
discovery of widget-specific methods within them difficult. The insert
method of the text widget is one such example. It will only get worse as
more methods are being implemented. Finding TextWidget.Insert after this
change is hopefully easier.

Only _examples/calc{,2}.go had to be changed in a single line, the
change is reasonably backward compatible. However, we are not bound to
keeping it while we are still at v0 tags anyway.